追加機能６実装完了
