--- ex01/fly_coukaton.py
@@ -1,5 +1,0 @@
-import sys
-import pygame as pg
-
-def main():
-    pg.display.set_mode((800, 600))
