+++ fig/beam.png
