+++ fig/explosion.gif
